+++ ups_mqtt/domain/exceptions.py
@@ -1,0 +1,10 @@
+class UpsDomainError(Exception):
+    pass
+
+
+class NutUnavailable(UpsDomainError):
+    pass
+
+
+class NutParseError(UpsDomainError):
+    pass